Change "int i" to "size_t i" in some for loops within the WX_PURE directives.
git-svn-id: https://svn.code.sf.net/p/fmslogo/source@3540 9d0050d7-00b0-40f9-871c-8e02cbd5794f

+++ trunk/logo/src/logodata.cpp
@@ -345,7 +345,7 @@ capital(
 char *cap_strnzcpy(char *dst, const char * src, int len)
 {
 #ifdef WX_PURE
-    for (int i = 0; i < len; i++)
+    for (size_t i = 0; i < len; i++)
     {
         dst[i] = capital(src[i]);
     }
@@ -385,7 +385,7 @@ char *noparitylow_strnzcpy(char *dst, const char *src, int len)
 int low_strncmp(const char *string1, const char * string2, int length)
 {
 #ifdef WX_PURE
-    for (int i = 0; i < length; i++)
+    for (size_t i = 0; i < length; i++)
     {
         if (*string1 != *string2)
         {
